+++ packages/function-swap/src/swap.ts
@@ -1,22 +1,38 @@
-import { recompose } from '@bemedev/decompose';
+import { getByKey, recompose } from '@bemedev/decompose';
 import { buildMap } from './helpers';
 import type {
   AnyFunction,
   FunctionSwap,
-  FunctionSwapObject,
+  FunctionSwapParams,
 } from './types';
 
 const _swap: FunctionSwap = fn => {
-  return ((...map: any[]) => {
+  const out: any = (...map: any[]) => {
     return (...newArgs: any[]) => {
       const decomposedMap = buildMap(map, newArgs);
       const recomposedArgs = recompose(decomposedMap) as any;
       return fn(...recomposedArgs);
     };
-  }) as any;
+  };
+
+  out.constraint = () => (keysMatch: Record<string, string>) => {
+    return (...newArgs: any[]) => {
+      const decomposedMap: Record<string, any> = {};
+
+      for (const [keyP, keyF] of Object.entries(keysMatch)) {
+        decomposedMap[keyF] = getByKey(newArgs, keyP);
+      }
+
+      const recomposedArgs = recompose(decomposedMap) as any;
+      return fn(...recomposedArgs);
+    };
+  };
+
+  return out;
 };
 
 export const swap = <const T extends AnyFunction>(fn: T) => _swap(fn);
 swap.fromFunction = _swap;
 swap.fromFn = _swap;
-swap.fromObject = (() => () => b => b) as FunctionSwapObject;
+swap.fromParameters = (() => () => b => b) as FunctionSwapParams;
+swap.fromParams = swap.fromParameters;