+++ vitest.config.ts
@@ -1,9 +1,7 @@
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-  server: {
-    host: '0.0.0.0',
-  },
+  server: { host: '0.0.0.0' },
 
   test: {
     passWithNoTests: true,
@@ -13,7 +11,7 @@ export default defineConfig({
       enabled: true,
       reportsDirectory: '.coverage',
       provider: 'v8',
-      exclude: [],
+      exclude: ['*.js', '*.cjs', '*.mjs'],
     },
 
     projects: [